Apply a hack to check for a truthy value
Ansible has trouble checking for truthy values of single variables. When
generating the Jinja template to check the condition, it uses the
variable's value instead of the variable itself. When the variable
contains a string, this can cause syntax errors. By using the `default`
filter, Ansible gets tricked into using the variable.